Fix handling of submodules that use relative URLs.
Pass the URL of the parent (sub)module into handle_submodules, so the
correct URL can be constructed in the case of relative URLs, e.g. as
used by modular Boost. For the git definition of relative URLs, see the
description under the 'add' command here:
https://www.kernel.org/pub/software/scm/git/docs/git-submodule.html

This can also be done with git submodule--helper resolve-relative-url
"$repo", but that would require a git upgrade to version 2.8.2 or
higher, and use git's internal api.