Neither the notify or build processes should (or do) use the webhook payload.
The trigger build functionality in the Admin interface was hitting an
issue where the lsr.payload was null in some situations. Rather than
working around it I decided to remove the payload from the real
build/notify processes. Note the payload is still present in the
handle_commit process since it is needed for relaying.

Signed-off-by: David Greaves <david.greaves@jolla.com>