Determine the obs and user outside of the loop.
Add some activity noise for the logs.

This also fixes a very strange error in this snippet:
            mapobj = WebHookMapping(
                repourl=repourl, branch=branch,
                user=User.objects.get(id=1),
                obs=BuildService.objects.all()[0],
                notify=False, build=False,
                project=project, package=package,
                placeholder=True,
            )

Where having the parameter:
     obs=BuildService.objects.all()[0],
failed but enclosing it in parentheses:
     obs=(BuildService.objects.all()[0]),
worked.

Clearly this should not happen and in fact it didn't happen in the
django shell. However it *did* happen in the participant code.

Signed-off-by: David Greaves <david.greaves@jolla.com>