The _read_conf() method is called repeatedly with different data sources
Each call used to use the data source or the default value. This meant
that if a value was set from one source but then not set in a
subsequent source the default value would be used.

Adding a check for an existing attr before falling back to the default
avoids this.

Signed-off-by: David Greaves <david.greaves@jolla.com>