The _read_conf() method is called repeatedly with different data sources
Each call used to use the data source or the default value. This meant
that if a value was set from one source but then not set in a
subsequent source the default value would be used.

The defaults have now been removed and the check order is explicit.

An addition check for missing config values is also made.

Signed-off-by: David Greaves <david.greaves@jolla.com>