Workaround for inconsistent repository submodule fetching.
handle_submodules function relies on .gitmodules file for fetching the
submodules. However if a repository has no submodules BUT still contains
.gitmodules file, git submodule update will fail with error "pathspec
foo did not match any file(s) known to git". As a workaround if
submodule update fails check from the git-dir whether the submodule
really exists in the repository or not and in case the repository is not
listed in git-dir config continue with next submodule.